[feat] #42 スライドキャンバス統合
